chore: comment mutations 분리
